# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orderServices": {
            "get": {
                "tags": [
                    "DonHangDichVu"
                ],
                "summary": "Lấy danh sách đơn hàng",
                "description": "Trả về danh sách tất cả các đơn hàng.",
                "operationId": "266204e4df7345409e3f9a7d43d8934f",
                "responses": {
                    "200": {
                        "description": "Dữ liệu được lấy thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dữ liệu được lấy thành công"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ServiceOrderResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "fail"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thông báo lỗi"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "DonHangDichVu"
                ],
                "summary": "Tạo đơn hàng mới",
                "description": "Tạo một đơn hàng mới cùng với chi tiết đơn hàng.",
                "operationId": "091bf1ad730b086c164d976f154f7e94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Mataikhoan",
                                    "NgayGiao",
                                    "chi_tiet"
                                ],
                                "properties": {
                                    "Mataikhoan": {
                                        "description": "Mã tài khoản của người đặt hàng",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "GhiChu": {
                                        "description": "Ghi chú cho đơn hàng",
                                        "type": "string",
                                        "example": "Ghi chú đơn hàng"
                                    },
                                    "NgayGiao": {
                                        "description": "Ngày giờ sử dụng đơn hàng",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-11-30 15:30:00"
                                    },
                                    "chi_tiet": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "MaSP": {
                                                    "description": "Mã dịch vụ",
                                                    "type": "integer",
                                                    "example": 60
                                                },
                                                "SoLuong": {
                                                    "description": "Số lượng dịch vụ",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thêm thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thêm thành công"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ServiceOrderResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Lỗi xác thực hoặc dữ liệu không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "fail"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thông báo lỗi"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orderServices/{Mataikhoan}": {
            "get": {
                "tags": [
                    "DonHangDichVu"
                ],
                "summary": "Lấy đơn hàng theo tài khoản",
                "description": "Trả về đơn hàng",
                "operationId": "5386e12e6771384e283aaf71a80cb800",
                "parameters": [
                    {
                        "name": "Mataikhoan",
                        "in": "path",
                        "description": "ID của tài khoản cần lấy thông tin đơn hàng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy dữ liệu thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy dữ liệu thành công"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ServiceOrderResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Lỗi khi lấy dữ liệu"
                    },
                    "404": {
                        "description": "Tài khoản không tìm thấy"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ServiceOrderResource": {
                "properties": {
                    "MaDonHang": {
                        "type": "integer",
                        "example": 1
                    },
                    "Mataikhoan": {
                        "type": "integer",
                        "example": 1
                    },
                    "TongTien": {
                        "type": "integer",
                        "example": 500000
                    },
                    "SoLuong": {
                        "type": "integer",
                        "example": 1
                    },
                    "Ten": {
                        "type": "string",
                        "example": "Nguyễn Văn A"
                    },
                    "SDT": {
                        "type": "string",
                        "example": "0123456789"
                    },
                    "DiaChi": {
                        "type": "string",
                        "example": "123 Đường ABC, Quận 1, TP.HCM"
                    },
                    "PTTT": {
                        "type": "string",
                        "example": "Chuyển khoản"
                    },
                    "GhiChu": {
                        "type": "string",
                        "example": "Ghi chú đơn hàng 1"
                    },
                    "Loai": {
                        "type": "string",
                        "example": "1"
                    },
                    "NgayDat": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-29T17:09:04Z"
                    },
                    "NgayGiao": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-01T17:09:04Z"
                    },
                    "TrangThai": {
                        "type": "string",
                        "example": "Đang xử lý"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-29T17:09:04Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-29T17:09:04Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}